Clean up iOS ffi.rs
This makes it easier to transition to a future autogenerated version of UIKit.
