+++ src/platform_impl/ios/uikit/device.rs
@@ -1,8 +1,7 @@
-use icrate::Foundation::{MainThreadMarker, NSObject};
+use icrate::Foundation::{MainThreadMarker, NSInteger, NSObject};
+use objc2::encode::{Encode, Encoding};
 use objc2::rc::Id;
 use objc2::{extern_class, extern_methods, msg_send_id, mutability, ClassType};
-
-use super::super::ffi::UIUserInterfaceIdiom;
 
 extern_class!(
     #[derive(Debug, PartialEq, Eq, Hash)]
@@ -24,3 +23,19 @@ extern_methods!(
         pub fn userInterfaceIdiom(&self) -> UIUserInterfaceIdiom;
     }
 );
+
+#[repr(transparent)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct UIUserInterfaceIdiom(NSInteger);
+
+unsafe impl Encode for UIUserInterfaceIdiom {
+    const ENCODING: Encoding = NSInteger::ENCODING;
+}
+
+impl UIUserInterfaceIdiom {
+    pub const Unspecified: UIUserInterfaceIdiom = UIUserInterfaceIdiom(-1);
+    pub const Phone: UIUserInterfaceIdiom = UIUserInterfaceIdiom(0);
+    pub const Pad: UIUserInterfaceIdiom = UIUserInterfaceIdiom(1);
+    pub const TV: UIUserInterfaceIdiom = UIUserInterfaceIdiom(2);
+    pub const CarPlay: UIUserInterfaceIdiom = UIUserInterfaceIdiom(3);
+}
