+++ src/platform_impl/ios/uikit/status_bar_style.rs
@@ -1,4 +1,3 @@
-use crate::platform::ios::StatusBarStyle;
 use icrate::Foundation::NSInteger;
 use objc2::encode::{Encode, Encoding};
 
@@ -10,16 +9,6 @@ pub enum UIStatusBarStyle {
     Default = 0,
     LightContent = 1,
     DarkContent = 3,
-}
-
-impl From<StatusBarStyle> for UIStatusBarStyle {
-    fn from(value: StatusBarStyle) -> Self {
-        match value {
-            StatusBarStyle::Default => Self::Default,
-            StatusBarStyle::LightContent => Self::LightContent,
-            StatusBarStyle::DarkContent => Self::DarkContent,
-        }
-    }
 }
 
 unsafe impl Encode for UIStatusBarStyle {
